back-end/requirements.txt
fastapi
uvicorn
sqlalchemy
chromadb
google-generativeai
python-dotenv
pytest
httpx
pydantic